--- controllers/users/getUserProfile.js
@@ -1,32 +1,0 @@
-const { dataSource } = require("../../db/data-source");
-const { validateSignup } = require("../../utils/validUtils");
-const AppError = require("../../utils/appError");
-const ERROR_MESSAGES = require("../../utils/errorMessages");
-
-async function getUserProfile(req, res, next) {
-  console.log(req.user);
-  const { id: userId } = req.user;
-  const user = await dataSource.getRepository("Users").findOne({
-    select: [
-      "name",
-      "email",
-      "phone",
-      "gender",
-      "birth_date",
-      "address_zipcode",
-      "address_district",
-      "address_detail",
-      "photo",
-    ],
-    where: { id: userId },
-  });
-
-  res.status(200).json({
-    status: true,
-    data: {
-      user,
-    },
-  });
-}
-
-module.exports = getUserProfile;
